audio bug fixed
